import ObjectViewer_Enhanced_20100110_src.zip 20100110 (2010-01-12)
This is an enhanced version of the OpenBVE tool Object Viewer. It is
based on the version 1.2.3.1 source code release.

Enhancements:
 * Added a button on the messages window to copy the messages to the
   clipboard.
 * Implemented the status icons in the message list, as per Route
   Viewer.
 * Fixed the Type column width in the message window.
 * Fixed a number of spelling errors.

I first wrote these enhancements over 7 months ago. At that time there
was no way to export the message log from the programme. I added an
option to copy the messages to the clipboard so that I could export them
for further examination. Since then, the author has added a function to
export to a file but having both options is even nicer.

Run the executable file from the OpenBVE directory as it needs access to
the data directory.

Enjoy.